requirements.txt
fastapi
uvicorn
llama-cpp-python==0.2.80